Modularized functions being called on resolves from each route
