+++ public/services/index.js
@@ -9,12 +9,12 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import 'plugins/wazuh/services/error-handler.js';
-import 'plugins/wazuh/services/theming.js';
-import 'plugins/wazuh/services/api-request.js';
-import 'plugins/wazuh/services/generic-request';
-import 'plugins/wazuh/services/data-handler';
-import 'plugins/wazuh/services/app-state';
-import 'plugins/wazuh/services/api-tester';
-import 'plugins/wazuh/services/pattern-handler';
-import 'plugins/wazuh/services/routes';
+import 'plugins/wazuh/services/error-handler'
+import 'plugins/wazuh/services/theming';
+import 'plugins/wazuh/services/api-request'
+import 'plugins/wazuh/services/generic-request'
+import 'plugins/wazuh/services/data-handler'
+import 'plugins/wazuh/services/app-state'
+import 'plugins/wazuh/services/api-tester'
+import 'plugins/wazuh/services/pattern-handler'
+import 'plugins/wazuh/services/routes'
