+++ public/services/resolves/go-to-kibana.js
@@ -1,0 +1,26 @@
+/*
+ * Wazuh app - File for routes definition
+ * Copyright (C) 2018 Wazuh, Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Find more information about this on the LICENSE file.
+ */
+// Manage leaving the app to another Kibana tab
+export default ($location, $window) => {
+
+    let url = $location.$$absUrl.substring(0, $location.$$absUrl.indexOf('#'));
+
+    if ($window.sessionStorage.getItem(`lastSubUrl:${url}`).includes('/wazuh#/visualize') ||
+        $window.sessionStorage.getItem(`lastSubUrl:${url}`).includes('/wazuh#/doc') ||
+        $window.sessionStorage.getItem(`lastSubUrl:${url}`).includes('/wazuh#/context')){
+
+            $window.sessionStorage.setItem(`lastSubUrl:${url}`, url);
+
+    }
+
+    $window.location.href = $location.absUrl().replace('/wazuh#', '/kibana#');
+}
