+++ public/services/resolves/health-check.js
@@ -1,0 +1,19 @@
+/*
+ * Wazuh app - File for routes definition
+ * Copyright (C) 2018 Wazuh, Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Find more information about this on the LICENSE file.
+ */
+export default ($window, $rootScope) => {
+    if (!$window.sessionStorage.getItem('healthCheck')) { // New session, execute health check
+        $window.sessionStorage.setItem('healthCheck', 'executed');
+        return true;
+    } else {
+        return false;
+    }
+}
